requirements.txt
pandas==1.0.1

dnspython
ipwhois

pytest